refactor: rename files to match kebab case
